+++ lib/clients/stremio_addon.py
@@ -3,7 +3,46 @@ from lib.utils.client_utils import show_dialog
 from lib.utils.utils import USER_AGENT_HEADER, IndexerType, info_hash_to_magnet
 from lib.stremio.addons_manager import Addon
 from lib.stremio.stream import Stream
+from lib.api.jacktook.kodi import kodilog
 
+
+
+class StremioAddonCatalogsClient(BaseClient):
+    def __init__(self, params):
+        super().__init__(None, None)
+        self.params = params
+        self.base_url = self.params["addon_url"]
+
+    def search(self, imdb_id, mode, media_type, season, episode, dialog):
+        pass
+
+    def parse_response(self, res):
+        pass
+
+    def get_catalog_info(self, skip, force_refresh=False):
+        url = f"{self.base_url}/catalog/{self.params['catalog_type']}/{self.params['catalog_id']}/skip={skip}.json"
+        kodilog(url)
+        res = self.session.get(url, headers=USER_AGENT_HEADER, timeout=10)
+        if res.status_code != 200:
+            return
+        return res.json()
+    
+    def get_meta_info(self):
+        url = f"{self.base_url}/meta/{self.params['catalog_type']}/{self.params['video_id']}.json"
+        kodilog(url)
+        res = self.session.get(url, headers=USER_AGENT_HEADER, timeout=10)
+        if res.status_code != 200:
+            return
+        return res.json()
+    
+    def get_stream_info(self):
+        url = f"{self.base_url}/stream/{self.params['catalog_type']}/{self.params['video_id']}.json"
+        kodilog(url)
+        res = self.session.get(url, headers=USER_AGENT_HEADER, timeout=10)
+        if res.status_code != 200:
+            return
+        return res.json()
+    
 
 class StremioAddonClient(BaseClient):
     def __init__(self, addon: Addon):