+++ lib/gui/resolver_window.py
@@ -108,6 +108,8 @@ class ResolverWindow(BaseWindow):
                 "tv_data": self.item_information["tv_data"],
             }
         )
+        if self.playback_info["is_pack"]:
+            self.resolve_pack()
 
     def resolve_pack(self):
         self.pack_data = get_pack_info(