+++ lib/router.py
@@ -45,10 +45,12 @@ from lib.navigation import (
     trakt_auth_revoke,
     trakt_list_content,
     tv_episodes_details,
+    tv_menu,
     tv_seasons_details,
     tv_shows_items,
 )
 from lib.search import search
+from lib.stremio.catalogs import list_stremio_catalog, list_stremio_episodes, list_stremio_seasons, list_stremio_tv, list_stremio_tv_streams
 from lib.telegram import (
     get_telegram_files,
     get_telegram_latest,
@@ -61,7 +63,7 @@ from lib.utils.torrent_utils import (
     torrent_action,
     torrent_files,
 )
-from lib.stremio.ui import stremio_login, stremio_toggle_addons, stremio_logout
+from lib.stremio.ui import stremio_login, stremio_toggle_addons, stremio_logout, stremio_toggle_catalogs, stremio_update
 
 def addon_router():
     param_string = sys.argv[2][1:]
@@ -123,8 +125,16 @@ def addon_router():
         "animation_menu": animation_menu,
         "animation_item": animation_item,
         "stremio_toggle_addons": stremio_toggle_addons,
+        "stremio_toggle_catalogs": stremio_toggle_catalogs,
+        "list_stremio_catalog": list_stremio_catalog,
+        "list_stremio_seasons": list_stremio_seasons,
+        "list_stremio_episodes": list_stremio_episodes,
+        "list_stremio_tv_streams": list_stremio_tv_streams,
+        "list_stremio_tv": list_stremio_tv,
+        "tv_menu": tv_menu,
         'stremio_login': stremio_login,
         'stremio_logout': stremio_logout,
+        "stremio_update": stremio_update,
     }
 
     if param_string: