+++ lib/stremio/ui.py
@@ -5,9 +5,10 @@ from lib.api.jacktook.kodi import kodilog
 from lib.db.cached import cache
 from datetime import timedelta
 from lib.stremio.client import Stremio
-from lib.utils.kodi_utils import ADDON
+from lib.utils.kodi_utils import ADDON, get_setting, set_setting
 
 STREMIO_ADDONS_KEY = "stremio_addons"
+STREMIO_CATALOGS_ADDONS_KEY = "stremio_catalog_addons"
 STREMIO_CATALOG_KEY = "stremio_catalog"
 
 
@@ -40,34 +41,60 @@ def get_selected_addon_urls() -> List[str]:
     return selected_addons.split(",")
 
 
+def get_selected_catalogs_addon_urls() -> List[str]:
+    selected_addons = cache.get(STREMIO_CATALOGS_ADDONS_KEY, hashed_key=True) or ""
+    return selected_addons.split(",")
+
+
 def get_selected_addons() -> List[Addon]:
     catalog = get_addons_catalog()
     selected_ids = cache.get(STREMIO_ADDONS_KEY, hashed_key=True) or ""
     return [addon for addon in catalog.addons if addon.key() in selected_ids]
 
 
-def stremio_login(params):
-    # Create a dialog box
-    dialog = xbmcgui.Dialog()
+def get_selected_catalogs_addons() -> List[Addon]:
+    catalog = get_addons_catalog()
+    selected_ids = cache.get(STREMIO_CATALOGS_ADDONS_KEY, hashed_key=True) or ""
+    return [addon for addon in catalog.addons if addon.key() in selected_ids]
 
+
+def stremio_login(params):
+    dialog = xbmcgui.Dialog()
     dialog.ok(
         "Stremio Add-ons Import",
         "To import your add-ons, please log in with your Stremio email and password.\n\n"
         + "Your login details will not be saved and are only used once for this process.",
     )
 
-    # Show an input dialog for email
     email = dialog.input(heading="Enter your Email", type=xbmcgui.INPUT_ALPHANUM)
-
     if not email:
         return
 
-    # Show a password dialog
     password = dialog.input(heading="Enter your Password", type=xbmcgui.INPUT_ALPHANUM)
-
     if not password:
         return
 
+    log_in(email, password, dialog)
+
+
+def stremio_update(params):
+    dialog = xbmcgui.Dialog()
+    confirm = dialog.yesno(
+        "Update Stremio Addons",
+        "Do you want to update the Addons from you account?",
+        nolabel="Cancel",
+        yeslabel="Yes",
+    )
+    if not confirm:
+        return
+
+    email = get_setting("stremio_email")
+    password = get_setting("stremio_pass")
+
+    log_in(email, password, dialog)
+
+
+def log_in(email, password, dialog):
     try:
         stremio = Stremio()
         stremio.login(email, password)
@@ -90,10 +117,27 @@ def stremio_login(params):
             timedelta(days=365 * 20),
             hashed_key=True,
         )
-        settings = ADDON.getSettings()
-        ADDON.setSetting("stremio_loggedin", "true")
-        settings.setString("stremio_email", email)
-        settings.setString("stremio_pass", password)
+
+        dialog = xbmcgui.Dialog()
+        confirm = dialog.yesno(
+            "Stremio Add-ons Import",
+            "Do you want also to import catalogs?",
+            nolabel="Cancel",
+            yeslabel="Yes",
+        )
+        if confirm:
+            catalogs = AddonManager(addons).get_addons_with_resource("catalog")
+            selected_catalogs = [catalog.key() for catalog in catalogs]
+            cache.set(
+                STREMIO_CATALOGS_ADDONS_KEY,
+                ",".join(selected_catalogs),
+                timedelta(days=365 * 20),
+                hashed_key=True,
+            )
+
+        set_setting("stremio_loggedin", "true")
+        set_setting("stremio_email", email)
+        set_setting("stremio_pass", password)
     except Exception as e:
         dialog.ok(
             "Add-ons Import Failed",
@@ -118,6 +162,7 @@ def stremio_logout(params):
     )
     if confirm:
         cache.set(STREMIO_ADDONS_KEY, None, timedelta(seconds=1), hashed_key=True)
+        cache.set(STREMIO_CATALOGS_ADDONS_KEY, None, timedelta(seconds=1), hashed_key=True)
         cache.set(STREMIO_CATALOG_KEY, None, timedelta(seconds=1), hashed_key=True)
         settings = ADDON.getSettings()
         ADDON.setSetting("stremio_loggedin", "false")
@@ -125,6 +170,52 @@ def stremio_logout(params):
         settings.setString("stremio_pass", "")
         _ = get_addons_catalog()
         stremio_toggle_addons(None)
+        stremio_toggle_catalogs(None)
+
+
+def stremio_toggle_catalogs(params):
+    kodilog("stremio_toggle_catalogs")
+    selected_ids = get_selected_catalogs_addon_urls()
+
+    addon_manager = get_addons_catalog()
+    addons = addon_manager.get_addons_with_resource("catalog")
+
+    dialog = xbmcgui.Dialog()
+    selected_addon_ids = [
+        addons.index(addon) for addon in addons if addon.key() in selected_ids
+    ]
+
+    options = []
+    for addon in addons:
+        option = xbmcgui.ListItem(
+            label=addon.manifest.name, label2=f"{addon.manifest.description}"
+        )
+
+        logo = addon.manifest.logo
+        if not logo or logo.endswith(".svg"):
+            logo = "DefaultAddon.png"
+
+        option.setArt({"icon": logo})
+        options.append(option)
+
+    settings = ADDON.getSettings()
+    stremio_email = settings.getString("stremio_email")
+    title = stremio_email or "Stremio Community Catalogs List"
+    selected_indexes = dialog.multiselect(
+        title, options, preselect=selected_addon_ids, useDetails=True
+    )
+
+    if selected_indexes is None:
+        return
+
+    selected_addon_ids = [addons[index].key() for index in selected_indexes]
+
+    cache.set(
+        STREMIO_CATALOGS_ADDONS_KEY,
+        ",".join(selected_addon_ids),
+        timedelta(days=365 * 20),
+        hashed_key=True,
+    )
 
 
 def stremio_toggle_addons(params):